dbd/extra/mongo/pyproject.toml
[tool.poetry]
name = "mongo"
version = "0.1.0"
description = ""
authors = ["Mutestock <doublenegative@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^3.11.3"
selenium = "^3.141.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
